Extract account general info into Account component
